pyproject.toml
[project]
name = "daily-idea-bot"
version = "0.1.0"
description = "AI Agent that suggests daily project ideas based on tech trends"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot>=21.0",
    "google-genai>=1.0.0",
    "httpx>=0.27.0",
    "apscheduler>=3.10.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "pytest>=8.0.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
